// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  READER: 'READER',
  AUTHOR: 'AUTHOR',
  MODERATOR: 'MODERATOR'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const PostStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED'
} as const

export type PostStatus = (typeof PostStatus)[keyof typeof PostStatus]